Move optimizer/executor tests to new file test_capi/test_opt.py
